pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "SilicaAnimus"
version = "0.2.0"
description = "Le bot du serveur Discord de la voie du Thalos"

readme = "README.md"
requires-python = ">=3.8"
keywords = ["discord", "HelloAsso"]
authors = [
  {name = "Patator", email = "lmpatator@gmail.com" }
]

classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
  "discord.py",
  "typer",
  "python-dotenv",
  "pytest",
  "pytest-asyncio",
  "pre-commit",
  "google-api-python-client",
  "unidecode>=1.4.0",
  "black>=24.8.0",
]

[project.scripts]
SilicaAnimus = "SilicaAnimus:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["src/tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
test = [
    "pytest>=8.3.5",
]
